more fixes...

+++ AuthTestClientCmd/Clients.h
@@ -40,7 +40,7 @@ public:
 	SecPkgContext_Sizes SecPkgContextSizes;
 
 	//Error handling
-	LONG	dwErrorCode;
+	DWORD	dwErrorCode;
 	WCHAR	szErrorLocation[255];
 	WCHAR	szErrorMessage[255];
 
@@ -62,7 +62,7 @@ private:
 	BOOL ReceiveBytes(SOCKET s, PBYTE pBuf, DWORD cbBuf, DWORD *pcbRead);
 	BOOL ReceiveAuthResult(int * iAuthResult);
 
-	void LogError(LONG dwError, LPCWSTR pszErrorLocation);
+	void LogError(DWORD dwError, LPCWSTR pszErrorLocation);
 
 	void PrintHexDump(DWORD length, PBYTE buffer);
